+++ src/shared/domain/exceptions/value-objects/invalid-filename/invalid-filename.exception.ts
@@ -1,0 +1,11 @@
+import { BaseException } from '@/shared/domain/exceptions/base.exception';
+
+/**
+ * Invalid Filename Exception
+ * This exception is thrown when a filename value is invalid.
+ */
+export class InvalidFilenameException extends BaseException {
+  constructor(message: string) {
+    super(message);
+  }
+}